+++ src/main/java/com/opsfactor/community/capability/masterdata/production/billofmaterials/facade/dto/ListaTecnicaComponenteDTO.java
@@ -13,6 +13,6 @@ public class ListaTecnicaComponenteDTO {
         
     private String componentMaterialUnitOfMeasureId;
     
-    private Float quantity;
+    private Double quantity;
     
 }
